+++ src/pydosert/__init__.py
@@ -1,3 +1,4 @@
+from pydosert.engine.photon_base_engine import PhotonBaseEngine
 from pydosert.engine.dose_engine import DoseEngine
 from pydosert.data import MachineConfig, OptimizationConfig, Phantom, Patient, Beam, BeamSequence
 from importlib.metadata import version, PackageNotFoundError
@@ -6,11 +7,12 @@ try:
     __version__ = version("pydosert")
 except PackageNotFoundError:
     __version__ = "0.0.0"
-    
-__all__ = ['DoseEngine', 
-           'MachineConfig', 
-           'OptimizationConfig', 
-           'Phantom', 
-           'Patient', 
-           'Beam', 
+
+__all__ = ['PhotonBaseEngine',
+           'DoseEngine',
+           'MachineConfig',
+           'OptimizationConfig',
+           'Phantom',
+           'Patient',
+           'Beam',
            'BeamSequence']